+++ orbus_interface/include/configurator/MotorPIDConfigurator.h
@@ -29,7 +29,9 @@
 *
 */
 
-#include "async_serial/ParserPacket.h"
+//#include "async_serial/ParserPacket.h"
+#include <lib_orb_cpp/include/PacketSerial.h>
+
 
 #include <ros/ros.h>
