+++ orbus_interface/lib_orb_cpp/src/ParserPacket.cpp
@@ -13,7 +13,7 @@
  * Lesser General Public License for more details.
  */
 
-#include "async_serial/ParserPacket.h"
+#include "ParserPacket.h"
 
 using namespace std;
 using namespace boost;
